+++ PowerEditor/src/WinControls/TaskList/TaskList.h
@@ -50,7 +50,7 @@ public:
 	virtual ~TaskList() {};
 	void init(HINSTANCE hInst, HWND hwnd, HIMAGELIST hImaLst, int nbItem, int index2set);
 	virtual void destroy();
-	void setFont(TCHAR *fontName, size_t fontSize);
+	void setFont(TCHAR *fontName, int fontSize);
 	RECT adjustSize();
 	int getCurrentIndex() const {return _currentIndex;}
 	int updateCurrentIndex();
